refactor(chatroom): relocate orchestration strategies under session-group (phase 5)
Move ConcurrentStrategy, SequentialStrategy, GroupChatStrategy, HandoffStrategy, MagenticStrategy, BaseStrategy, the legacy OrchestrationStrategy/OrchestrationContext types, and createStrategy from packages/services/src/chatroom/orchestration/ to packages/services/src/session-group/orchestrators/. Strategy bodies and their ~76KB of tests are unchanged — git tracked the renames so blame is preserved.

Drop the chatroom/orchestration/stream-agent.ts re-export shim added in phase 2; relocated strategies now import streamAgentTurn / sendToAgentWithRetry from the sibling stream-session.ts directly.

chatroom/orchestration/ now contains only product-layer concerns the chatroom owns: shared.ts (XML/JSON helpers used by ChatroomService.buildPrompt), approval-gate.ts, observability.ts. The chatroom barrel re-exports OrchestrationStrategy/OrchestrationContext/createStrategy from session-group for back-compat with existing callers.

wrapStrategy stays as the documented bridge from the legacy OrchestrationStrategy contract to SessionGroupOrchestrator. Rewriting strategies to consume SessionGroupRunContext directly (and deleting wrapStrategy) is a clean follow-up PR.

README updated. All 1112 vitest specs pass; npm run lint clean.
