+++ packages/services/src/chatroom/index.ts
@@ -1,5 +1,5 @@
 export { ChatroomService, type ChatroomSessionFactory } from './ChatroomService';
-export { createStrategy } from './orchestration';
+export { createStrategy } from '../session-group';
 export { ApprovalGate } from './orchestration';
-export type { OrchestrationStrategy, OrchestrationContext } from './orchestration';
+export type { OrchestrationStrategy, OrchestrationContext } from '../session-group';
 export type { ApprovalGateConfig, ApprovalGateResult, ApprovalHandler, ApprovalLogEntry } from './orchestration';
