+++ packages/services/src/session-group/observability.test.ts
