+++ packages/services/src/session-group/orchestrators/index.ts
@@ -1,4 +1,15 @@
 export { wrapStrategy } from './wrap-strategy';
+export { createStrategy } from './createStrategy';
+export { ConcurrentStrategy } from './ConcurrentStrategy';
+export { SequentialStrategy } from './SequentialStrategy';
+export { GroupChatStrategy } from './GroupChatStrategy';
+export { HandoffStrategy } from './HandoffStrategy';
+export { MagenticStrategy } from './MagenticStrategy';
+export { BaseStrategy } from './legacy-types';
+export type {
+  OrchestrationStrategy,
+  OrchestrationContext,
+} from './legacy-types';
 export type {
   ProductHooks,
   SessionGroupOrchestrator,
