+++ packages/services/src/session-group/shared.ts
